requirements.txt
discord.py==2.4.0
logging==0.4.9.6
sortedcontainers==2.4.0
